Avoid the use of stride tricks to make code more readable
Admittedly this is a style issue, but I feel like stridetricks isn't
very useful in this particular case and it somewhat destroys readability
for me.

While the reshaping is cute, I honestly don't think it helps much with
the python overhead.

I hit this code in trying to implment frustum culling for my application
(though I still haven't succeeded in getting a performance improvement).

Let me know what you think.

I think there is also some complexity in the code due to an attempt at
making it generaizable to beyond a single "aabb" object.

However, I didn't see any tests for this, and so my presumption is that
this is problem and likely the cause of much of the original code
compelxity.